intercom_sync/requirements.txt
requests>=2.31.0
supabase>=2.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.9.0
tenacity>=8.2.3
pydantic>=2.5.0
rich>=13.7.0
click>=8.1.7
pytz>=2024.1